[Silver III] Title: 칸토어 집합, Time: 4 ms, Memory: 4260 KB -BaekjoonHub

+++ 백준/Silver/4779. 칸토어 집합/칸토어 집합.cc
@@ -3,7 +3,7 @@ using namespace std;
 
 vector<string> dp(13, "\0");
 
-string cantor(const long long &n) {
+string cantor(const int &n) {
     if (dp[n] != "\0") return dp[n];
     string tmp1 = cantor(n-1), tmp2 = "";
     for (int i = 0; i < tmp1.length(); i++) tmp2 += " ";
@@ -12,7 +12,7 @@ string cantor(const long long &n) {
 
 int main() {
     ios::sync_with_stdio(0); cin.tie(0);
-    long long n;
+    int n;
     dp[0] = "-";
     while (cin >> n) cout << cantor(n) << '\n';
 }
